+++ src/backend/core/datatypes/BigInt2DoubleFilter.h
@@ -31,7 +31,7 @@ public:
 	}
 
 	//! Return the data type of the column
-	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Numeric; }
+	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Double; }
 
 protected:
 	//! Using typed ports: only bigint inputs are accepted
