+++ src/backend/core/datatypes/DayOfWeek2DoubleFilter.h
@@ -29,7 +29,7 @@ public:
 	}
 
 	//! Return the data type of the column
-	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Numeric; }
+	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Double; }
 
 protected:
 	//! Using typed ports: only date-time inputs are accepted.
