+++ src/backend/core/datatypes/String2DoubleFilter.h
@@ -41,7 +41,7 @@ public:
 	}
 
 	//! Return the data type of the column
-	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Numeric; }
+	AbstractColumn::ColumnMode columnMode() const override { return AbstractColumn::ColumnMode::Double; }
 
 protected:
 	//! Using typed ports: only string inputs are accepted.
